# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Tutume E-mzigo",
        "description": "Tutume E-mzigo"
    },
    "paths": {
        "/customer/signup": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "Signup",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Signup",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.Signup"
            }
        },
        "/customer/update-profile": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "Update",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.Update"
            }
        },
        "/customer/signin": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "Signin",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Signin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.Signin"
            }
        },
        "/customer/get-profile": {
            "get": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "GetProfile",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "customer_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "CustomerController.GetProfile"
            }
        },
        "/customer/logout-delete": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "LogoutDelete",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LogoutDelete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": ["logout", "delete"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.LogoutDelete"
            }
        },
        "/customer/change-password": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "ChangePassword",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ChangePassword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.ChangePassword"
            }
        },
        "/customer/check-price": {
            "get": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "CheckPrice",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CheckPrice",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "customer_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "vehicle_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "delivery_option_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "pickup_address",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "delivery_address",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "CustomerController.CheckPrice"
            }
        },
...
        "/customer/apply-coupon": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "ApplyCoupon",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ApplyCoupon",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "coupon_code": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.ApplyCoupon"
            }
        },
...
        "/customer/cancel-job": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "CancelJob",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "CancelJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.CancelJob"
            }
        },
        "/customer/recently-shipped": {
            "get": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "RecentlyShipped",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "RecentlyShipped",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "customer_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "CustomerController.RecentlyShipped"
            }
        },
        "/customer/track-job": {
            "get": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "TrackJob",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "TrackJob",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "customer_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "job_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "CustomerController.TrackJob"
            }
        },
...
        "/provider/logout-delete": {
            "post": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "LogoutDelete",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LogoutDelete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "provider_id": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": ["logout", "delete"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProviderController.LogoutDelete"
            }
        },
...
    },
...
    "servers": [{
        "url": "http://localhost:1440/api"
    }]
}